+++ test/index.js
@@ -8,16 +8,6 @@ const errorHandler = require('./errorHandler.test')
 const execute = require('./execute.test')
 const isEnumerableObject = require('./isEnumerableObject.test')
 
-process
-  .on('uncaughtException', error => {
-    console.error('Uncaught exception:', error)
-    process.exitCode = 1
-  })
-  .on('unhandledRejection', error => {
-    console.error('Unhandled rejection:', error)
-    process.exitCode = 1
-  })
-
 const tests = new TestDirector()
 
 isEnumerableObject(tests)